refactor auth and me routes + tests
